Fixed rounding to integer in number of points

+++ pulse_lib/keysight/M3202A_uploader.py
@@ -404,7 +404,7 @@ class UploadAggregator:
             # work with sample rate in GSa/s
             sample_rate = (seg.sample_rate if seg.sample_rate is not None else job.default_sample_rate) * 1e-9
             duration = seg.total_time[tuple(job.index)]
-            npt =  round(duration * sample_rate)
+            npt =  int(round(duration * sample_rate))
             info = SegmentRenderInfo(sample_rate, t_start, npt)
             logging.info(f'duration: {duration}, sample_rate: {sample_rate}, segment:{info}')
             segments.append(info)
